Time: 245 ms (97.10%), Space: 58.2 MB (43.66%) - LeetHub

+++ 1711-count-good-meals/1711-count-good-meals.cpp
@@ -22,7 +22,7 @@ public:
     int countPairs(vector<int>& deliciousness) {
         unordered_map<int, int>mp;
         vector<int>b(33, true);
-        unsigned long long ret = 0;
+        unsigned int ret = 0;
         for(int x : deliciousness) {
             ++mp[x];
         }
